sensors: roll `sensor-types` into `sensor-api` #(#1683)
Now that `host-sp-messages` no longer depends on `task-sensor-types`,
there's really no reason for it to be a separate crate. Thus, let's
just move the error types into `task-sensor-api`.
