--- task/sensor-types/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "task-sensor-types"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-hubpack.workspace = true
-num-derive.workspace = true
-num-traits.workspace = true
-serde.workspace = true
-zerocopy.workspace = true
-
-counters = { path = "../../lib/counters" }
-derive-idol-err.path = "../../lib/derive-idol-err"
-drv-i2c-types.path = "../../drv/i2c-types"
